Move ClassHierarchy out of internal and into parser
